+++ jkit-sql/src/main/java/com/alianga/jkit/sql/SqlTokenType.java
@@ -278,6 +278,7 @@ public enum SqlTokenType {
     DECLARE,
     VACUUM,
     COPY,
+    LOAD,
     HANDLER,
     PREPARE,
     EXECUTE,